Store prover stdout and stderr in temporary files
This changes the way Benchpress collects the prover's outputs from using
a pipe to using temporary files that are read once the prover is done.
On Unix systems, the temporary files default to being stored in
`/dev/shm`; on other platforms, the default temporary directory is used.

This ensures that the prover never gets stuck waiting for benchpress to
read from pipes, but more importantly minimises context switches between
the prover and benchpress.